requirements.txt
# 核心计算库
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 硬件接口
pyserial>=3.5

# 优化求解器
cvxpy>=1.2.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0
pytest-xdist>=2.5.0
pytest-timeout>=2.1.0
pytest-rerunfailures>=10.0

# 代码质量
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0
mypy>=0.990

# 数据分析（用于Notebook）
pandas>=1.4.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.7.0
